requirements.txt
pandas==2.0.3
numpy==1.24.3
flask==2.3.2
pyecharts==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.1
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.9.0
python-dotenv==1.0.0